pyproject.toml
[project]
name = "locallora"
version = "0.1.0"
requires-python = ">=3.11,<3.13"
dependencies = [
    "pymupdf>=1.24.0",
    "torch==2.5.1",
    "torchao==0.6.1",
    "transformers>=4.46.0,<4.49.0",
    "datasets",
    "accelerate>=0.34.0",
    "peft>=0.13.0",
    "trl>=0.12.0",
    "tensorboard",
    "httpx",
    "arxiv",
    "ads",
    "semanticscholar",
    "pyyaml",
    "sentencepiece",
    "protobuf",
    "mlx-lm>=0.19.0",
]

[tool.black]
line-length = 100
